Update .gitignore and remove cached files
