+++ .gitignore
@@ -21,4 +21,6 @@ Thumbs.db
 # Vite
 vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
+
+#Python Notebooks
 *.ipynb
