+++ requirements.txt
@@ -1,0 +1,4 @@
+fastapi
+pandas
+playsound
+requests
